Ink needs a ground: one mechanism replacing the same colour bug found four times
A control drew itself against the *page* while sitting on the selection band, and
came out unreadable on any palette whose band is bright. It was fixed locally four
separate times — the switch that became a black dot on white, the chip that became
a pill-shaped hole, the card that drew zero pixels different from the band, the
row whose dim second line vanished.

`Theme::ground` is those four corrections said once. A container tells its children
what they are being drawn on (`Ground::{Page, Band, Chrome}`), `Ink::resolve` reads
it, and three hand-written `if selected` sites were deleted with the parity sheets
still reporting `identical`. `Ink::Chrome` and `Ink::Selection` deliberately ignore
it: they name a surface, not a role on one.

Also here:

- `Palette::error`, per palette and derived in `from_seed`. Luma on this handset is
  green-heavy (54r + 183g + 19b), so pure red measures 53 — an error colour has to
  be pink on dark and brick on light, and the guard test holds every palette to a
  70-luma delta on all three grounds.
- `Tile`, the 45th component: a rounded square with one letter, squared and centred
  in `draw` and not only in `measure`, because `CrossAlign::Stretch` hands it the
  whole band whatever it measured.
- `Widget::focus_state`, so a row can ask its control whether it has focus instead
  of being told separately and drifting.

The catalogue is documented in `docs/ui-catalog.md`, which is its index and its
contract.

+++ crates/symbian-ui/src/paint.rs
@@ -207,6 +207,10 @@ pub fn band_round(c: &mut Canvas, r: Rect, s: &Surface, radius: i32) {
 ///
 /// A flat fill and not a [`Surface`]: this sits behind text at body size, where a gradient reads as
 /// dirt on the screen rather than as depth.
+// Nine parameters, and they are one text run: where it starts, how tall it is, what it says, which
+// part of it is picked, in what font, in what colour. A struct would be built at both call sites and
+// taken apart again here — a type whose whole job is to satisfy a counter.
+#[allow(clippy::too_many_arguments)]
 pub fn text_selection(
     c: &mut Canvas,
     x: i32,
@@ -454,7 +458,7 @@ mod tests {
                     Color::hex(0xFFFFFF),
                 );
             }
-            let top = (0..40).find(|&y| px(&buf, size, 1, y) == thumb).unwrap() as i32;
+            let top = (0..40).find(|&y| px(&buf, size, 1, y) == thumb).unwrap();
             assert!(top >= last_top, "offset {offset} moved the thumb up");
             last_top = top;
         }
